Instead of creating a new serializer, let's reuse the already existing "IntOrStringSerializer"
