+++ common/src/main/kotlin/entity/DiscordMessage.kt
@@ -4,6 +4,7 @@ import dev.kord.common.entity.optional.Optional
 import dev.kord.common.entity.optional.OptionalBoolean
 import dev.kord.common.entity.optional.OptionalInt
 import dev.kord.common.entity.optional.OptionalSnowflake
+import dev.kord.common.serialization.IntOrStringSerializer
 import kotlinx.datetime.Instant
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
@@ -96,7 +97,7 @@ public data class DiscordMessage(
     val attachments: List<DiscordAttachment>,
     val embeds: List<DiscordEmbed>,
     val reactions: Optional<List<Reaction>> = Optional.Missing(),
-    val nonce: Optional<@Serializable(with = NonceAsStringSerializer::class) String> = Optional.Missing(),
+    val nonce: Optional<@Serializable(with = IntOrStringSerializer::class) String> = Optional.Missing(),
     val pinned: Boolean,
     @SerialName("webhook_id")
     val webhookId: OptionalSnowflake = OptionalSnowflake.Missing,
@@ -120,26 +121,6 @@ public data class DiscordMessage(
     val interaction: Optional<DiscordMessageInteraction> = Optional.Missing(),
     val thread: Optional<DiscordChannel> = Optional.Missing()
 )
-
-internal object NonceAsStringSerializer : KSerializer<String> {
-    override val descriptor: SerialDescriptor
-        get() = PrimitiveSerialDescriptor("nonce", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): String {
-        val jsonDecoder = decoder as? JsonDecoder ?: error("Can be deserialized only by JSON")
-        val element = jsonDecoder.decodeJsonElement()
-            .jsonPrimitive
-
-        return when {
-            element.jsonPrimitive.isString -> element.content
-            else -> element.int.toString()
-        }
-    }
-
-    override fun serialize(encoder: Encoder, value: String) {
-        encoder.encodeString(value)
-    }
-}
 
 /**
  * @param id id of the sticker
